src/api-server/pyproject.toml
# LOCKFILES: Generate with `pip-compile pyproject.toml -o requirements.lock`
#   For dev: `pip-compile pyproject.toml --extra dev -o requirements-dev.lock`
#   Install locked: `pip install -r requirements.lock`

[project]
name = "directai-api-server"
version = "0.1.0"
description = "DirectAI — OpenAI-compatible inference API routing layer"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "httpx[http2]>=0.27,<1",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.5,<3",
    "pyyaml>=6.0,<7",
    "python-multipart>=0.0.12",
    "prometheus-client>=0.21,<1",
    "aiosqlite>=0.20,<1",
    "opentelemetry-api>=1.27,<2",
    "opentelemetry-sdk>=1.27,<2",
    "opentelemetry-instrumentation-fastapi>=0.48b0,<1",
    "opentelemetry-instrumentation-httpx>=0.48b0,<1",
    "azure-monitor-opentelemetry-exporter>=1.0.0b30,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx",             # TestClient uses httpx
    "ruff>=0.8,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 120
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
